+++ tests/CSharpToJs.Tests/AttributeTests.cs
@@ -1,74 +1,40 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Reflection;
-using System.Text;
-using CSharpToJs.Core.Attributes;
+﻿using CSharpToJs.Core.Attributes;
 using CSharpToJs.Core.Interfaces;
 using CSharpToJs.Core.Models;
 using CSharpToJs.Core.Services;
 using CSharpToJs.Tests.Dummies;
 using CSharpToJs.Tests.Mocks;
+using System;
+using System.Linq;
+using System.Reflection;
 using Xunit;
 
 namespace CSharpToJs.Tests
 {
     public class AttributeTests
     {
+        
+
         [Fact]
-        public void JsPropertyConverterAttribute()
+        public void JsClassConverterAttribute_ThrowsIfNotAClassConverter()
         {
-            var mock = new CustomPropertyConverterClass();
-            var propInfo = mock.GetType().GetProperty(nameof(mock.MyProperty));
-            var converterContext = new PropertyConverterContext();
+            Action action = () => new JsClassConverterAttribute(typeof(string));
 
-            // The mock always returns "Custom": "Super" as the name/value
-            var (expectedName, expectedValue) = ("Custom", "Super");
-            var attribute = propInfo.GetCustomAttribute<JsPropertyConverterAttribute>();
-            var customConverter = attribute.PropertyConverter;
-            var converterInstance = Activator.CreateInstance(customConverter.GetType()) as IJsPropertyConverter;
-            var propResult = converterInstance?.Convert(converterContext);
+            Assert.Throws<ArgumentException>(action);
+        }
 
+        [Fact]
+        public void PropertyResolver_ThrowsIfNotAPropertyResolver()
+        {
+            Action action = () => new PropertyResolverAttribute(typeof(string));
 
-            Assert.NotNull(customConverter);
-            Assert.NotNull(propResult);
-            Assert.Equal(expectedName, propResult.Name);
-            Assert.Equal(expectedValue, propResult.Value);
+            Assert.Throws<ArgumentException>(action);
         }
 
         [Fact]
         public void JsPropertyConverterAttribute_ThrowsIfNotAPropertyConverter()
         {
             Action action = () => new JsPropertyConverterAttribute(typeof(string));
-
-            Assert.Throws<ArgumentException>(action);
-        }
-
-        [Fact]
-        public void DefaultPropertyResolverIgnoreAttribute()
-        {
-            var propertyResolver = new PropertyResolver();
-            var shouldNotContain = nameof(ClassDummy.IShouldBeIgnored);
-
-            var props = propertyResolver.GetProperties(typeof(ClassDummy)).ToList();
-
-            Assert.DoesNotContain(props, a => a.Name == shouldNotContain);
-        }
-
-        [Fact]
-        public void JsClassConverterAttribute()
-        {
-            var resolver = new ClassConverterResolver();
-
-            var converter = resolver.Resolve(typeof(CustomClassConverterDummy));
-
-            Assert.IsType<CustomClassConverterMock>(converter);
-        }
-
-        [Fact]
-        public void JsClassConverterAttribute_ThrowsIfNotAClassConverter()
-        {
-            Action action = () => new JsClassConverterAttribute(typeof(string));
 
             Assert.Throws<ArgumentException>(action);
         }
