+++ tests/CSharpToJs.Tests/ServiceTests.cs
@@ -4,6 +4,8 @@ using System.IO;
 using System.Linq;
 using System.Reflection;
 using System.Text;
+using CSharpToJs.Core.Attributes;
+using CSharpToJs.Core.Interfaces;
 using CSharpToJs.Core.Models;
 using CSharpToJs.Core.Services;
 using CSharpToJs.Tests.Dummies;
@@ -144,5 +146,48 @@ namespace CSharpToJs.Tests
 
             Assert.IsType<JsClassConverter>(resolver.DefaultClassConverter);
         }
+
+        [Fact]
+        public void CustomJsPropertyConverterThroughAttribute()
+        {
+            var mock = new CustomPropertyConverterClass();
+            var propInfo = mock.GetType().GetProperty(nameof(mock.MyProperty));
+            var converterContext = new PropertyConverterContext();
+
+            // The mock always returns "Custom": "Super" as the name/value
+            var (expectedName, expectedValue) = ("Custom", "Super");
+            var attribute = propInfo.GetCustomAttribute<JsPropertyConverterAttribute>();
+            var customConverter = attribute.PropertyConverter;
+            var converterInstance = Activator.CreateInstance(customConverter.GetType()) as IJsPropertyConverter;
+            var propResult = converterInstance?.Convert(converterContext);
+
+
+            Assert.NotNull(customConverter);
+            Assert.NotNull(propResult);
+            Assert.Equal(expectedName, propResult.Name);
+            Assert.Equal(expectedValue, propResult.Value);
+        }
+
+        [Fact]
+        public void DefaultPropertyResolverIgnoreAttribute()
+        {
+            var propertyResolver = new PropertyResolver();
+            var shouldNotContain = nameof(ClassDummy.IShouldBeIgnored);
+
+            var props = propertyResolver.GetProperties(typeof(ClassDummy)).ToList();
+
+            Assert.DoesNotContain(props, a => a.Name == shouldNotContain);
+        }
+
+        [Fact]
+        public void ClassConverterResolverForAttribute()
+        {
+            var resolver = new ClassConverterResolver();
+
+            var converter = resolver.Resolve(typeof(CustomClassConverterDummy));
+
+            Assert.IsType<CustomClassConverterMock>(converter);
+        }
+
     }
 }
